refactor(app): orgnize context and state into seperate files
