+++ apps/web/app/multiplayer/ydoc-context.tsx
@@ -2,10 +2,13 @@
 
 import { toUint8Array } from "js-base64";
 import React, { createContext, useEffect, useMemo, useState } from "react";
+import { bind } from "valtio-yjs";
+
 import { WebrtcProvider } from "y-webrtc";
 import * as Y from "yjs";
+import { multiplayerState } from "./multiplayer-state";
 
-const YDocContext = createContext({
+const multiplayerContext = createContext({
   ydoc: new Y.Doc(),
   provider: undefined as undefined | WebrtcProvider,
 });
@@ -39,11 +42,16 @@ export const YDocProvider = ({
     }
   }, [yDocUpdate, ydoc]);
 
+  useEffect(() => {
+    const unbind = bind(multiplayerState, ydoc.getMap("multiplayer-state"));
+    return () => unbind();
+  }, [ydoc]);
+
   return (
-    <YDocContext.Provider value={{ ydoc, provider }}>
+    <multiplayerContext.Provider value={{ ydoc, provider }}>
       {children}
-    </YDocContext.Provider>
+    </multiplayerContext.Provider>
   );
 };
 
-export const useYDoc = () => React.useContext(YDocContext);
+export const useYDoc = () => React.useContext(multiplayerContext);
